+++ contracts/protocol/tokenization/StableDebtToken.sol
@@ -43,7 +43,7 @@ contract StableDebtToken is IStableDebtToken, DebtTokenBase {
     function initialize(
         ILendingPool pool,
         address underlyingAsset,
-        uint16 trancheId,
+        uint64 trancheId,
         IAaveIncentivesController incentivesController,
         uint8 debtTokenDecimals,
         string memory debtTokenName,
